pyproject.toml
[tool.poetry]
name = "futuremind project data warehouse"
version = "0.1.0"
description = "Data warehous project"
authors = ["Radoslaw Zamojski <radoslaw.zamojski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.3.0"
requests = "^2.32.4"
sqlalchemy = "^2.0.41"
oracledb = "^3.1.1"
python-dotenv = "^1.1.0"
matplotlib = "^3.10.3"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
